% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_viz.R
\name{knn_viz}
\alias{knn_viz}
\title{Visualize the kNN classifiers}
\usage{
knn_viz(train, cl, k = 7, validate = NULL, cl_validate = NULL,
  prob_size = T)
}
\arguments{
\item{train}{matrix or data.frame with 2 columns}

\item{cl}{vector of training classes}

\item{k}{number of neighbors}

\item{validate}{matrix or data.frame with 2 columns}

\item{cl_validate}{vector of classes}

\item{prob_size}{whether to adjust the size of points depending on the estimated probabilities}
}
\value{
a ggplot object
}
\description{
Plots the kNN decision boundary for 2D problems.
}
