% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.split.R
\name{prop.split}
\alias{prop.split}
\title{Function to split data}
\usage{
prop.split(N, p)
}
\arguments{
\item{N}{number of points}

\item{p}{vector of proportions. Must sum 1.}
}
\value{
A list of \code{length(p)} vectors, with sample indexes from \code{1} to \code{N}.
}
\description{
Split \code{N} points in proportions given by \code{p}
}
