% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_viz.R
\name{svm_viz}
\alias{svm_viz}
\title{Visualize the SVM classifiers}
\usage{
svm_viz(train, cl, kernel = "linear")
}
\arguments{
\item{train}{matrix or data.frame with 2 columns}

\item{cl}{vector of training classes}

\item{kernel}{the kernel. See \code{?e1071::svm}}
}
\value{
a ggplot object
}
\description{
Plots the SVM decision boundary for 2D problems.
}
